# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Harency ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/send_otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send Otp",
                "description": "Send otp to verify user",
                "operationId": "authSendOtp",
                "requestBody": {
                    "description": "Pass user phone number and type if its 0 or 1,0 for verify user and 1 for forgot password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "type"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "format": "phone_number",
                                        "example": "9786543210"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "type",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong Phone number response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone number. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/social_media_login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login with Social Media",
                "description": "Login with social media :- Facebook, Instagram, Twitter, Google ",
                "operationId": "authSocialLogin",
                "requestBody": {
                    "description": "Pass user phone number, name, email, social media id, social media type <br><br>social_media_type:<br> 1 for facebook, <br> 2 for google, <br> 3 for instagram, <br>  4 for twitter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "social_id",
                                    "social_media_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "abc"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "abc@example.com"
                                    },
                                    "social_id": {
                                        "type": "string",
                                        "example": "kjh435h43kj5"
                                    },
                                    "social_media_type": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "format": "device_token",
                                        "example": "1232443bcvhsdgdfs"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "format": "device_type",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}